Work has begun on correcting the logic of the check_ip module. Created a separate project branch.
Combining this branch with the main branch is expected after the program as
a whole will work properly and without errors.

At the moment, the module works, but the necessary changes have not been made
to the main module so that the program can run.

Validation and normalization of data entering the class is transferred to the
control of the decorator. Getters and setters were also added, the task of
which is to transfer data to the class for further processing.

Some methods were removed from the module, the function of which was taken
over by the decorator.

Type hinting has been set for the new methods as well as for the decorator.

===========================================================================

On branch bugfix/ip_check
Changes to be committed:
	modified:   check_ip.py