Finalization of the module for testing check_ip.py
Finish writing tests for this module.
Add the date and time of the test to the detailed report.
Translate all keys from return '1' to return 'details' and the like.
Create a basis for transferring a report to a file using the 'set_note' key.

On branch bugfix/ip_check
Changes to be committed:
	modified:   test_check_ip.py

Changes not staged for commit:
	modified:   ../ip_checker/check_ip.py
	modified:   ../poetry.lock
	modified:   ../pyproject.toml