Fix the work of the test_check_ip.py module
Make it possible to write test report logs to a file by key.
Create a control method for all tests and set up the logic for executing a single text by key there.
Fix single testing.
Fix a bug with mass testing.

On branch bugfix/ip_check
Changes to be committed:
	modified:   test_check_ip.py

Changes not staged for commit:
	modified:   ../ip_checker/check_ip.py
	modified:   ../poetry.lock
	modified:   ../pyproject.toml